backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.1
python-multipart==0.0.12
librosa==0.10.2
numpy==1.26.4
scikit-learn==1.5.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyJWT>=2.10.1
